# specs/002-physical-ai-book/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Physical AI & Humanoid Robotics Textbook API
  description: |
    FastAPI backend for Physical AI textbook with RAG chatbot, conversation persistence,
    and user authentication. Supports textbook content ingestion, question answering (RAG and selected-text modes),
    user signup/signin, and reusable agent skills.
  version: 1.0.0
  contact:
    name: Panaversity Hackathon
    url: https://panaversity.org

servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://physical-ai-backend.onrender.com
    description: Production server (example)

tags:
  - name: health
    description: Health check endpoints
  - name: ingestion
    description: Textbook content ingestion
  - name: rag
    description: RAG question answering
  - name: auth
    description: User authentication
  - name: profile
    description: User profile management
  - name: skills
    description: Reusable agent skills

paths:
  /ping:
    get:
      summary: Health check
      description: Verify API is running
      tags:
        - health
      responses:
        '200':
          description: API is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "ok"
                  message:
                    type: string
                    example: "Physical AI Textbook API is running"

  /ingest:
    post:
      summary: Ingest textbook content
      description: |
        Reads all textbook markdown files, chunks content (200-300 tokens with 50-token overlap),
        creates embeddings using OpenAI text-embedding-3-small, and stores in Qdrant Cloud.
      tags:
        - ingestion
      responses:
        '200':
          description: Ingestion successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: "success"
                  chapters_processed:
                    type: integer
                    example: 20
                  total_chunks:
                    type: integer
                    example: 850
                  collection_name:
                    type: string
                    example: "physical_ai_textbook"
        '500':
          description: Ingestion failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /ask:
    post:
      summary: Ask question (RAG mode)
      description: |
        Answer question using RAG: query Qdrant for relevant chunks, generate answer with OpenAI ChatKit SDK,
        return answer with source citations, and save conversation to Neon Postgres.
      tags:
        - rag
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  minLength: 3
                  maxLength: 2000
                  example: "What is Physical AI?"
                session_id:
                  type: string
                  description: Optional session ID for anonymous users
                  example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Answer generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskResponse'
        '400':
          description: Invalid request (empty question, etc.)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (invalid JWT token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error (Qdrant/OpenAI/Postgres failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /ask_selected:
    post:
      summary: Ask question about selected text
      description: |
        Answer question using only provided selected text as context (no RAG query).
        Save conversation to Neon Postgres with question_type='selected_text'.
      tags:
        - rag
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
                - context
              properties:
                question:
                  type: string
                  minLength: 3
                  maxLength: 2000
                  example: "Explain this concept in simple terms"
                context:
                  type: string
                  minLength: 10
                  maxLength: 5000
                  description: Selected text from textbook
                  example: "ROS 2 is a middleware framework for robot control..."
                session_id:
                  type: string
                  description: Optional session ID for anonymous users
                  example: "550e8400-e29b-41d4-a716-446655440000"
      responses:
        '200':
          description: Answer generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AskSelectedResponse'
        '400':
          description: Invalid request (empty question/context, context too short)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized (invalid JWT token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error (OpenAI/Postgres failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/signup:
    post:
      summary: User signup
      description: |
        Create new user account with email/password and background information.
        Hash password with bcrypt, store in Neon Postgres, generate JWT token.
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SignupRequest'
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request (email already exists, password too short)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error (Postgres failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/signin:
    post:
      summary: User signin
      description: |
        Authenticate user with email/password, verify password hash, generate JWT token,
        update last_login timestamp.
      tags:
        - auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                  format: email
                  example: "student@university.edu"
                password:
                  type: string
                  format: password
                  example: "password123"
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Authentication failed (invalid email or password)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error (Postgres failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /auth/me:
    get:
      summary: Get current user
      description: Get authenticated user's profile from JWT token
      tags:
        - auth
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized (invalid or expired token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /profile:
    get:
      summary: Get user profile
      description: Get authenticated user's full profile including backgrounds
      tags:
        - profile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          description: Unauthorized (invalid or expired token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error (Postgres failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /skills/summarize:
    post:
      summary: Summarize textbook section
      description: |
        Use SummarizeSection agent skill (Persona + Questions + Principles pattern)
        to generate concise summary of textbook content.
      tags:
        - skills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Textbook content to summarize
                  example: "ROS 2 is a middleware framework for robot control. It provides communication between processes via topics, services, and actions..."
                max_length:
                  type: integer
                  description: Maximum summary length in words
                  default: 200
                  example: 150
      responses:
        '200':
          description: Summary generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: string
                    example: "ROS 2 provides middleware for robot control through topics, services, and actions..."
                  original_length:
                    type: integer
                    example: 450
                  summary_length:
                    type: integer
                    example: 120
                  compression_ratio:
                    type: number
                    format: float
                    example: 0.27
                  model:
                    type: string
                    example: "gpt-3.5-turbo"
        '400':
          description: Invalid request (empty text)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error (OpenAI failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /skills/quiz:
    post:
      summary: Generate quiz questions
      description: |
        Use GenerateQuizQuestions agent skill to create multiple-choice questions
        for self-assessment based on textbook content.
      tags:
        - skills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - text
              properties:
                text:
                  type: string
                  description: Textbook content to create quiz from
                  example: "NVIDIA Isaac Sim is a photorealistic simulation platform built on Omniverse..."
                num_questions:
                  type: integer
                  description: Number of questions to generate
                  default: 3
                  minimum: 1
                  maximum: 10
                  example: 5
      responses:
        '200':
          description: Quiz questions generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  questions:
                    type: array
                    items:
                      type: object
                      properties:
                        question:
                          type: string
                          example: "What platform is Isaac Sim built on?"
                        options:
                          type: array
                          items:
                            type: string
                          example: ["Gazebo", "Unity", "Omniverse", "Unreal Engine"]
                        correct_answer:
                          type: integer
                          description: Index of correct answer (0-based)
                          example: 2
                        explanation:
                          type: string
                          example: "Isaac Sim is built on NVIDIA Omniverse for photorealistic rendering."
                  num_questions:
                    type: integer
                    example: 5
                  model:
                    type: string
                    example: "gpt-3.5-turbo"
        '400':
          description: Invalid request (empty text, invalid num_questions)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error (OpenAI failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /skills/explain:
    post:
      summary: Explain technical term
      description: |
        Use ExplainTerm agent skill to explain technical terms from the textbook.
        If no context provided, uses RAG to find relevant textbook sections.
      tags:
        - skills
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - term
              properties:
                term:
                  type: string
                  description: Technical term to explain
                  example: "VSLAM"
                context:
                  type: string
                  description: Optional context (if omitted, uses RAG to find relevant content)
                  example: "Visual Simultaneous Localization and Mapping is used in Isaac ROS..."
      responses:
        '200':
          description: Explanation generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  term:
                    type: string
                    example: "VSLAM"
                  explanation:
                    type: string
                    example: "VSLAM (Visual Simultaneous Localization and Mapping) is a technique where a robot uses camera data to simultaneously build a map of its environment and determine its position within that map..."
                  context_used:
                    type: string
                    description: Indicates if RAG was used or user-provided context
                    example: "rag"
                  model:
                    type: string
                    example: "gpt-3.5-turbo"
        '400':
          description: Invalid request (empty term)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Server error (OpenAI/Qdrant failure)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    Error:
      type: object
      properties:
        detail:
          type: string
          example: "Invalid request: question cannot be empty"
        error_type:
          type: string
          example: "ValidationError"

    AskResponse:
      type: object
      properties:
        answer_text:
          type: string
          example: "Physical AI refers to AI systems that operate in and interact with the physical world..."
        sources:
          type: array
          items:
            $ref: '#/components/schemas/Citation'
        confidence:
          type: number
          format: float
          description: Optional confidence score
          example: 0.85
        conversation_id:
          type: integer
          example: 42
        model:
          type: string
          example: "gpt-3.5-turbo"

    AskSelectedResponse:
      type: object
      properties:
        answer_text:
          type: string
          example: "This concept describes middleware for robot control..."
        sources:
          type: array
          items:
            type: object
            properties:
              source_type:
                type: string
                example: "user_selection"
              text_excerpt:
                type: string
                example: "ROS 2 is a middleware framework..."
        conversation_id:
          type: integer
          example: 43
        model:
          type: string
          example: "gpt-3.5-turbo"

    Citation:
      type: object
      properties:
        module:
          type: string
          example: "Module 1: ROS 2"
        chapter:
          type: string
          example: "ROS 2 Architecture"
        chunk_id:
          type: string
          example: "chunk_42"
        relevance_score:
          type: number
          format: float
          example: 0.89

    SignupRequest:
      type: object
      required:
        - email
        - password
        - software_background
        - hardware_background
      properties:
        email:
          type: string
          format: email
          example: "student@university.edu"
        password:
          type: string
          format: password
          minLength: 8
          example: "password123"
        software_background:
          $ref: '#/components/schemas/SoftwareBackground'
        hardware_background:
          $ref: '#/components/schemas/HardwareBackground'

    SoftwareBackground:
      type: object
      properties:
        programming_languages:
          type: array
          items:
            type: string
          example: ["Python", "C++", "JavaScript"]
        robotics_experience:
          type: string
          enum: ["none", "beginner", "intermediate", "advanced"]
          example: "intermediate"
        ai_ml_level:
          type: string
          enum: ["none", "basic", "intermediate", "advanced"]
          example: "basic"
        prior_courses:
          type: array
          items:
            type: string
          example: ["CS101", "Robotics 101"]

    HardwareBackground:
      type: object
      properties:
        rtx_gpu_access:
          type: boolean
          example: true
        rtx_gpu_model:
          type: string
          example: "RTX 4070 Ti"
        jetson_kit:
          type: string
          example: "Orin Nano 8GB"
        robot_hardware:
          type: string
          example: "none"

    AuthResponse:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "student@university.edu"
        token:
          type: string
          description: JWT token for authentication
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."

    UserProfile:
      type: object
      properties:
        user_id:
          type: integer
          example: 1
        email:
          type: string
          format: email
          example: "student@university.edu"
        software_background:
          $ref: '#/components/schemas/SoftwareBackground'
        hardware_background:
          $ref: '#/components/schemas/HardwareBackground'
        created_at:
          type: string
          format: date-time
          example: "2025-11-27T12:34:56.789Z"
        last_login:
          type: string
          format: date-time
          example: "2025-11-28T08:15:30.123Z"